Initial commit
Basic funcionality appears to be working. I removed from the c files
the functions to write and read from file. I will have to figure out
how to compile the C code and link it in the simplest and most robust way.

I might use GeometryBasics instead of the structs here.