Several changes, but removed outside dependencies
There are two packages that define several of the things used by this
package: GeometryBasics and Meshes. Since this package is so simple,
I decided to remove the GeometryBasics dependency and related functions.

The only thing that impacts it is the new `Vertex` struct.

Also implemented the method `centroid` to calculate the centroid
of `GPCPolygons`.